delete file from mediastore on ViewPagerActivity in the background

+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/ViewPagerActivity.kt
@@ -181,8 +181,10 @@ class ViewPagerActivity : SimpleActivity(), ViewPager.OnPageChangeListener, View
         }
 
         val file = File(mPath)
-        if (!file.exists()) {
-            deleteFromMediaStore(file)
+        if (!file.exists() && file.length() == 0L) {
+            Thread {
+                deleteFromMediaStore(file)
+            }.start()
             finish()
             return
         }
